backend/requirements/production.txt
Django==3.1.6d
djangorestframework==3.12.2
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
gunicorn==20.0.4
requests==2.25.1
requests-toolbelt==0.9.1